+++ railties/CHANGELOG.md
@@ -1,3 +1,6 @@
+* Generate `app/views/layouts/application.html.erb` template's `<html>` with
+  the `[lang]` attribute set to the value of `I18n.locale`.
 
+  *Sean Doyle*
 
 Please check [7-0-stable](https://github.com/rails/rails/blob/7-0-stable/railties/CHANGELOG.md) for previous changes.